--- a/base/abstractarray.jl
+++ b/base/abstractarray.jl
@@ -1017,13 +1017,13 @@ function copyto_unaliased!(deststyle::IndexStyle, dest::AbstractArray, srcstyle:
     if deststyle isa IndexLinear
         if srcstyle isa IndexLinear
             # Single-index implementation
-            @inbounds for i in srcinds
+            @inbounds @simd for i in srcinds
                 dest[i + Δi] = src[i]
             end
         else
             # Dual-index implementation
             i = idf - 1
-            @inbounds for a in src
+            @inbounds @simd for a in src
                 dest[i+=1] = a
             end
         end
@@ -1031,8 +1031,8 @@ function copyto_unaliased!(deststyle::IndexStyle, dest::AbstractArray, srcstyle:
         iterdest, itersrc = eachindex(dest), eachindex(src)
         if iterdest == itersrc
             # Shared-iterator implementation
-            for I in iterdest
-                @inbounds dest[I] = src[I]
+            @inbounds @simd for I in iterdest
+                dest[I] = src[I]
             end
         else
             # Dual-iterator implementation