As suggested here: https://discourse.julialang.org/t/deferred-shape-arrays/48177/10, this should speed up copying from `ReshapedArray`s.

Tests, before:
```
julia> A = rand(2000); B = zeros(100,20);

julia> @btime $B .= reshape($A, (100, 20));
  263.395 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  944.538 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  789.962 ns (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  635.908 ns (0 allocations: 0 bytes)
```
after:
```
julia> A = rand(2000); B = zeros(100,20);

julia> @btime $B .= reshape($A, (100, 20));
  291.971 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());  # much improved
  182.660 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  759.545 ns (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())'; # hmm, really?
  681.230 ns (0 allocations: 0 bytes)
```