@chriselrod 2020-10-14
For me, adding `ivdep` slightly improves the performance for the double adjoint versions.
Is it worth adding a check for both the src and destination being `adjoint` or `transpose` wrappers, and if so, use `eachindex(parent(src))` instead?