@mcabbott 2020-10-14
Thanks for looking. My guess is that both being transposed is super-rare, it was just the first way I thought of to hit the CartesianIndex-ed path. 

On my laptop today, this change seems to speed up *both* of the transposed copies, although I see no effect from changing to `@simd ivdep`. Before:
```
julia> @btime $B' .= reshape($A, (100, 20))';
  3.754 μs (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  3.942 μs (0 allocations: 0 bytes)
```
After:
```
julia> @btime $B' .= reshape($A, (100, 20))';
  1.791 μs (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  1.750 μs (0 allocations: 0 bytes)
```