@chriselrod 2020-10-14
This is what I get:
```julia
julia> A = rand(2000); B = zeros(100,20); # first, current master benchmarks

julia> @btime $B .= reshape($A, (100, 20));
  137.586 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  880.385 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  1.376 μs (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  1.215 μs (0 allocations: 0 bytes)

julia> @btime $B .= reshape($A, (100, 20));
  142.991 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  881.173 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  1.377 μs (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  1.212 μs (0 allocations: 0 bytes)

julia> function Base.copyto_unaliased!(deststyle::IndexStyle, dest::AbstractArray, srcstyle::IndexStyle, src::AbstractArray)
           isempty(src) && return dest
           destinds, srcinds = LinearIndices(dest), LinearIndices(src)
           idf, isf = first(destinds), first(srcinds)
           Δi = idf - isf
           (checkbounds(Bool, destinds, isf+Δi) & checkbounds(Bool, destinds, last(srcinds)+Δi)) ||
               throw(BoundsError(dest, srcinds))
           if deststyle isa IndexLinear
               if srcstyle isa IndexLinear
                   # Single-index implementation
                   @inbounds @simd for i in srcinds
                       dest[i + Δi] = src[i]
                   end
               else
                   # Dual-index implementation
                   i = idf - 1
                   @inbounds for a in src
                       dest[i+=1] = a
                   end
               end
           else
               iterdest, itersrc = eachindex(dest), eachindex(src)
               if iterdest == itersrc
                   # Shared-iterator implementation
                   @inbounds @simd for I in iterdest
                       dest[I] = src[I]
                   end
               else
                   # Dual-iterator implementation
                   ret = iterate(iterdest)
                   @inbounds for a in src
                       idx, state = ret
                       dest[idx] = a
                       ret = iterate(iterdest, state)
                   end
               end
           end
           return dest
       end

julia> @btime $B .= reshape($A, (100, 20));
  141.395 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  104.703 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  1.184 μs (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  1.212 μs (0 allocations: 0 bytes)

julia> @btime $B .= reshape($A, (100, 20));
  140.036 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  108.115 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  1.183 μs (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  1.212 μs (0 allocations: 0 bytes)

julia> function Base.copyto_unaliased!(deststyle::IndexStyle, dest::AbstractArray, srcstyle::IndexStyle, src::AbstractArray)
           isempty(src) && return dest
           destinds, srcinds = LinearIndices(dest), LinearIndices(src)
           idf, isf = first(destinds), first(srcinds)
           Δi = idf - isf
           (checkbounds(Bool, destinds, isf+Δi) & checkbounds(Bool, destinds, last(srcinds)+Δi)) ||
               throw(BoundsError(dest, srcinds))
           if deststyle isa IndexLinear
               if srcstyle isa IndexLinear
                   # Single-index implementation
                   @inbounds @simd ivdep for i in srcinds
                       dest[i + Δi] = src[i]
                   end
               else
                   # Dual-index implementation
                   i = idf - 1
                   @inbounds for a in src
                       dest[i+=1] = a
                   end
               end
           else
               iterdest, itersrc = eachindex(dest), eachindex(src)
               if iterdest == itersrc
                   # Shared-iterator implementation
                   @inbounds @simd ivdep for I in iterdest
                       dest[I] = src[I]
                   end
               else
                   # Dual-iterator implementation
                   ret = iterate(iterdest)
                   @inbounds for a in src
                       idx, state = ret
                       dest[idx] = a
                       ret = iterate(iterdest, state)
                   end
               end
           end
           return dest
       end

julia> @btime $B .= reshape($A, (100, 20));
  140.963 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  105.879 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  971.294 ns (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  920.600 ns (0 allocations: 0 bytes)

julia> @btime $B .= reshape($A, (100, 20));
  144.164 ns (1 allocation: 64 bytes)

julia> @btime $B .= Base.ReshapedArray($A, (100, 20), ());
  104.377 ns (0 allocations: 0 bytes)

julia> @btime $B' .= reshape($A, (100, 20))';
  970.941 ns (1 allocation: 64 bytes)

julia> @btime $B' .= Base.ReshapedArray($A, (100, 20), ())';
  920.886 ns (0 allocations: 0 bytes)
```

To look at ASM:
```julia
julia> Bt = B'; At = Base.ReshapedArray(A, (100,20), ())';

julia> @code_native debuginfo=:none Base.copyto_unaliased!(IndexStyle(typeof(Bt)), B', IndexStyle(typeof(At)), Base.ReshapedArray(A, (100,20), ())')
```
Main loop with `@simd ivdep`:
```asm
L624:
        vpmullq %zmm7, %zmm2, %zmm8
        vpaddq  %zmm8, %zmm6, %zmm8
        vpsllq  $3, %zmm8, %zmm8
        vpaddq  %zmm8, %zmm4, %zmm8
        kxnorw  %k0, %k0, %k1
        vpgatherqq      (,%zmm8), %zmm9 {%k1}
        vpmullq %zmm7, %zmm3, %zmm8
        vpaddq  %zmm8, %zmm6, %zmm8
        vpsllq  $3, %zmm8, %zmm8
        vpaddq  %zmm8, %zmm5, %zmm8
        kxnorw  %k0, %k0, %k1
        vpscatterqq     %zmm9, (,%zmm8) {%k1}
        vpaddq  %zmm1, %zmm7, %zmm7
        addq    $-8, %rdx
        jne     L624 # repeat inner
        cmpq    %rcx, 24(%rsp)
        je      L528 # repeat outer
```
with `@simd` (no ivdep):
```asm
L768:
        vpmullq %zmm7, %zmm2, %zmm8
        vpaddq  %zmm8, %zmm6, %zmm8
        vpsllq  $3, %zmm8, %zmm8
        vpaddq  %zmm8, %zmm4, %zmm8
        kxnorw  %k0, %k0, %k1
        vpgatherqq      (,%zmm8), %zmm9 {%k1}
        vpmullq %zmm7, %zmm3, %zmm8
        vpaddq  %zmm8, %zmm6, %zmm8
        vpsllq  $3, %zmm8, %zmm8
        vpaddq  %zmm8, %zmm5, %zmm8
        kxnorw  %k0, %k0, %k1
        vpscatterqq     %zmm9, (,%zmm8) {%k1}
        vpaddq  %zmm1, %zmm7, %zmm7
        addq    $-8, %rbx
        jne     L768 # repeat inner
        cmpq    %rsi, %r9
        movq    64(%rsp), %r10
        je      L544 # repeat outer
```
So the inner loops are identical, but the outer loop is more involved when it doesn't have `ivdep`.
Adding the `ivdep` reduces this:
```asm
L544:
        leaq    1(%r8), %rdx
        movabsq $9223372036854775800, %rsi      # imm = 0x7FFFFFFFFFFFFFF8
        addq    $7, %rsi
        cmpq    %r15, %r14
        cmovbq  72(%rsp), %rsi
        incq    %rcx
        cmpq    %rsi, %r8
        movq    %rdx, %r8
        movq    32(%rsp), %r11
        je      L1465
L591:
        cmpq    $8, %r11
        jae     L608
        xorl    %esi, %esi
        jmp     L880
        nopl    (%rax)
L608:
        leaq    (%r12,%rcx,8), %rbx
        movq    48(%rsp), %rdx
        leaq    (%rdx,%rcx), %rsi
        leaq    (%r12,%rsi,8), %rsi
        movq    56(%rsp), %rdx
        leaq    (%rdx,%rcx), %r9
        cmpq    %rsi, %rbx
        movq    %rsi, %rdx
        cmovaq  %rbx, %rdx
        cmovbq  %rbx, %rsi
        leaq    (,%r9,8), %rbx
        addq    %r13, %rbx
        leaq    (,%rcx,8), %r9
        addq    %r13, %r9
        cmpq    %r9, %rbx
        movq    %r9, %r10
        cmovbq  %rbx, %r10
        cmovaq  %rbx, %r9
        addq    $8, %r9
        cmpq    %r9, %rsi
        jae     L720
        addq    $8, %rdx
        cmpq    %rdx, %r10
        jae     L720
        xorl    %esi, %esi
        movq    64(%rsp), %r10
        jmp     L880
        nopw    (%rax,%rax)
L720:
```
into this:
```asm
L528:
        leaq    1(%r12), %rcx
        leaq    7(%r9), %rdx
        cmpq    %r15, %r14
        cmovbq  48(%rsp), %rdx
        incq    %r13
        cmpq    %rdx, %r12
        movq    %rcx, %r12
        je      L1325
L561:
        movq    24(%rsp), %rcx
        cmpq    $8, %rcx
        jae     L592
        xorl    %ecx, %ecx
        jmp     L727
        nopw    %cs:(%rax,%rax)
L592:
        leaq    -1(%r12), %rdx
        andq    %r9, %rcx
        vpbroadcastq    %rdx, %zmm6
        movq    %r8, %rdx
        vmovdqa64       %zmm0, %zmm7
        nopw    (%rax,%rax)
L624:
```

But it might be architecture dependent. I'm guessing that these checks that are skipped by adding `ivdep` are only needed when it is SIMD (checks aren't needed if evaluating serially), and SIMD without un-transposing requires AVX-512.